[ember] @ember/component as source of types, ember as re-export
