+++ types/ember-modal-dialog/tsconfig.json
@@ -15,7 +15,9 @@
         ],
         "paths": {
             "@ember/object": ["ember__object"],
-            "@ember/object/*": ["ember__object/*"]
+            "@ember/object/*": ["ember__object/*"],
+            "@ember/component": ["ember__component"],
+            "@ember/component/*": ["ember__component/*"]
         },
         "types": [],
         "noEmit": true,
