+++ types/ember/tsconfig.json
@@ -26,6 +26,8 @@
             "@ember/debug/*": ["ember__debug/*"],
             "@ember/object": ["ember__object"],
             "@ember/object/*": ["ember__object/*"],
+            "@ember/component": ["ember__component"],
+            "@ember/component/*": ["ember__component/*"],
             "@ember/controller": ["ember__controller"],
             "@ember/polyfills": ["ember__polyfills"]
         },
