+++ types/ember__component/-private/action-support.d.ts
@@ -1,0 +1,8 @@
+import Mixin from "@ember/object/mixin";
+
+interface ActionSupport {
+    sendAction(action: string, ...params: any[]): void;
+}
+declare const ActionSupport: Mixin<ActionSupport>;
+
+export default ActionSupport;
