+++ types/ember__component/checkbox.d.ts
@@ -1,3 +1,8 @@
 import Ember from 'ember';
+import Component from '@ember/component';
 
-export default class Checkbox extends Ember.Checkbox { }
+/**
+ * The internal class used to create text inputs when the {{input}} helper is used
+ * with type of checkbox. See Handlebars.helpers.input for usage details.
+ */
+export default class Checkbox extends Component {}
