+++ types/ember__component/helper.d.ts
@@ -1,6 +1,27 @@
-import Ember from 'ember';
+import EmberObject from "@ember/object";
 
-export default class Helper extends Ember.Helper { }
+/**
+ * Ember Helpers are functions that can compute values, and are used in templates.
+ * For example, this code calls a helper named `format-currency`:
+ */
+export default class Helper extends EmberObject {
+    /**
+     * In many cases, the ceremony of a full `Ember.Helper` class is not required.
+     * The `helper` method create pure-function helpers without instances. For
+     * example:
+     */
+    static helper(helper: (params: any[], hash?: object) => any): Helper;
+    /**
+     * Override this function when writing a class-based helper.
+     */
+    compute(params: any[], hash: object): any;
+    /**
+     * On a class-based helper, it may be useful to force a recomputation of that
+     * helpers value. This is akin to `rerender` on a component.
+     */
+    recompute(): any;
+}
+
 /**
  * In many cases, the ceremony of a full `Helper` class is not required.
  * The `helper` method create pure-function helpers without instances. For
