+++ types/ember__component/tsconfig.json
@@ -17,6 +17,8 @@
         "paths": {
             "@ember/object": ["ember__object"],
             "@ember/object/*": ["ember__object/*"],
+            "@ember/controller": ["ember__controller"],
+            "@ember/controller/*": ["ember__controller/*"],
             "@ember/component": ["ember__component"],
             "@ember/component/*": ["ember__component/*"]
         },
@@ -30,6 +32,12 @@
         "text-area.d.ts",
         "text-field.d.ts",
         "helper.d.ts",
+        "-private/core-view.d.ts",
+        "-private/class-names-support.d.ts",
+        "-private/view-mixin.d.ts",
+        "-private/action-support.d.ts",
+        "-private/target-action-support.d.ts",
+        "-private/text-support.d.ts",
         "test/lib/assert.ts",
         "test/component.ts",
         "test/helper.ts"
